+++ packages/common/web/src/Utils/contexts.jsx
@@ -1,3 +1,3 @@
 import React from 'react';
 
-export const isMobileContext = React.createContext(false);
+export const HelpersContext = React.createContext({ isMobile: false, t: () => {} });